% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{accel_plot}
\alias{accel_plot}
\title{Plot UKBiobank Accelerometry Data
This function plots UKBiobank accelerometry data.}
\usage{
accel_plot(x)
}
\arguments{
\item{x}{an object inherited from data.frame. It is
assumed to have a \code{time}, \code{X}, \code{Y}, and \code{Z} column along with
either a \code{time} or \code{freq} column.}
}
\value{
a time-series plot faceted by axis.
}
\description{
Plot UKBiobank Accelerometry Data
This function plots UKBiobank accelerometry data.
}
\examples{
data(ukb_accel)
accel_plot(ukb_accel[1:1000,])
}
