% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{The Template \code{hello} Function.}
\usage{
hello(name = NULL, invisible = FALSE)
}
\arguments{
\item{name}{the name of the person/thing to say hello to. Default is \code{NULL}
meaning no-one/-thing.}

\item{invisible}{should the return be invisible? Default is \code{FALSE} (no).}
}
\value{
A \code{character} vector saying hello.
}
\description{
Write hello to the console.
}
