% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukb_accel}
\alias{ukb_accel}
\title{Synthesized UKBiobank Accelerometry Data}
\description{
A toy accelerometry dataset from UKBiobank where:
}
\details{
\code{time} is the time of the measurement;
\code{X} acceleration in the "X" direction;
\code{Y} acceleration in the "Y" direction;
\code{Z} acceleration in the "Z" direction;
}
\keyword{data}
